backend/requirements.txt
# ── Web framework ─────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy==2.0.30

# ── News / scraping ───────────────────────────────────────────────────────────
feedparser==6.0.11
beautifulsoup4==4.12.3
requests==2.31.0

# ── LLM providers (install the one you need) ─────────────────────────────────
openai==1.30.1
google-generativeai==0.5.4

# ── Excel generation ──────────────────────────────────────────────────────────
openpyxl==3.1.2

# ── Scheduling ────────────────────────────────────────────────────────────────
apscheduler==3.10.4

# ── Utilities ────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
lxml==5.2.2
